Add files via upload
Tic Tac Toe, also known as Noughts and Crosses, is a two-player game where the objective is to form a line of three of one's symbols (either "X" or "O") horizontally, vertically, or diagonally on a 3x3 grid. Players take turns placing their symbols on the grid. The game ends when one player forms a line of three or the grid is full, resulting in a tie. It's a simple yet strategic game, often used as an introductory project for programming due to its straightforward rules and logic.